services/cart-management-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
alembic==1.13.1

pytest==8.0.0
httpx==0.26.0
ruff==0.2.0
